requirements.txt
asgiref==3.8.1
Django==5.0.7
sqlparse==0.5.1
whitenoise==6.8.2
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-filter==24.2
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.9.0
pyparsing==3.2.0
tzdata==2024.1